services/ingest/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
kafka-python==2.0.2
prometheus-client==0.19.0
python-json-logger==2.0.7

